// generated/models/SubjectsSubjectForSelection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Learning API
 * API with password OAuth2 authentication
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface SubjectsSubjectForSelection
 */
export interface SubjectsSubjectForSelection {
    /**
     *
     * @type {string}
     * @memberof SubjectsSubjectForSelection
     */
    displayName?: string;
    /**
     *
     * @type {string}
     * @memberof SubjectsSubjectForSelection
     */
    fullName?: string;
    /**
     *
     * @type {string}
     * @memberof SubjectsSubjectForSelection
     */
    id?: string;
    /**
     *
     * @type {string}
     * @memberof SubjectsSubjectForSelection
     */
    parentId?: string;
    /**
     *
     * @type {string}
     * @memberof SubjectsSubjectForSelection
     */
    url?: string;
}

/**
 * Check if a given object implements the SubjectsSubjectForSelection interface.
 */
export function instanceOfSubjectsSubjectForSelection(value: object): value is SubjectsSubjectForSelection {
    return true;
}

export function SubjectsSubjectForSelectionFromJSON(json: any): SubjectsSubjectForSelection {
    return SubjectsSubjectForSelectionFromJSONTyped(json, false);
}

export function SubjectsSubjectForSelectionFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubjectsSubjectForSelection {
    if (json == null) {
        return json;
    }
    return {
        displayName: json["display_name"] == null ? undefined : json["display_name"],
        fullName: json["full_name"] == null ? undefined : json["full_name"],
        id: json["id"] == null ? undefined : json["id"],
        parentId: json["parent_id"] == null ? undefined : json["parent_id"],
        url: json["url"] == null ? undefined : json["url"],
    };
}

export function SubjectsSubjectForSelectionToJSON(json: any): SubjectsSubjectForSelection {
    return SubjectsSubjectForSelectionToJSONTyped(json, false);
}

export function SubjectsSubjectForSelectionToJSONTyped(value?: SubjectsSubjectForSelection | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        display_name: value["displayName"],
        full_name: value["fullName"],
        id: value["id"],
        parent_id: value["parentId"],
        url: value["url"],
    };
}